% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-functions.R
\name{mean_plus}
\alias{mean_plus}
\title{Take the mean of a vector and add to it}
\usage{
mean_plus(x, plus)
}
\arguments{
\item{x}{A numeric vector to take the mean of}

\item{plus}{A number to add to the mean}
}
\description{
Take the mean of a vector and add to it
}
\examples{
mean_plus(rnorm(10), 2)
}
